Move Area to separate package
