+++ src/main/kotlin/world/player/wilderness/obelisk/Obelisk.kt
@@ -1,7 +1,7 @@
 package world.player.wilderness.obelisk
 
 import com.google.common.collect.ImmutableSet
-import io.luna.game.model.Area
+import io.luna.game.model.area.Area
 import io.luna.game.model.Position
 
 /**
